[chore] package 정리
