+++ app/src/main/java/org/sopt/sample/data/remote/request/RequestLogin.kt
@@ -1,12 +1,11 @@
-package org.sopt.sample.data.remote
+package org.sopt.sample.data.remote.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignup(
+data class RequestLogin(
     val email: String,
     @SerialName("password")
-    val pw: String,
-    val name: String
+    val pw: String
 )
