+++ app/src/main/java/org/sopt/sample/data/remote/request/RequestSignup.kt
@@ -1,11 +1,12 @@
-package org.sopt.sample.data.remote
+package org.sopt.sample.data.remote.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestLogin(
+data class RequestSignup(
     val email: String,
     @SerialName("password")
-    val pw: String
+    val pw: String,
+    val name: String
 )
