+++ app/src/main/java/org/sopt/sample/data/remote/response/ResponseLogin.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.data.remote
+package org.sopt.sample.data.remote.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
